requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.38
langserve[all]>=0.1.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
faiss-cpu>=1.8.0
pypdf>=4.0.0
python-dotenv>=1.0.0
